# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: FalseFlag Control Plane API
  description: |
    The FalseFlag control-plane REST surface. Slice 3 extends slice 2's
    projects+flags+evaluate surface with environments, segments,
    snapshots, audit search, and evaluation-with-trace. The full
    operator/CRD-integration surface lands in slice 4.
  version: 0.3.0
  contact:
    name: FalseFlag
    url: https://github.com/depot/falseflag
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/projects/{slug}/segments:
    parameters:
      - $ref: "#/components/parameters/ProjectSlug"
    get:
      operationId: listSegments
      summary: List segments in the project
      tags: [segments]
      responses:
        "200":
          description: Segment listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SegmentList"
        "404":
          $ref: "#/components/responses/Error"
    post:
      operationId: createSegment
      summary: Create a project-scoped reusable predicate
      tags: [segments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSegmentRequest"
      responses:
        "201":
          description: Segment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Segment"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /v1/projects/{slug}/segments/{segKey}:
    parameters:
      - $ref: "#/components/parameters/ProjectSlug"
      - $ref: "#/components/parameters/SegmentKey"
    get:
      operationId: getSegment
      summary: Get a segment by key
      tags: [segments]
      responses:
        "200":
          description: Segment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Segment"
        "404":
          $ref: "#/components/responses/Error"
    put:
      operationId: updateSegment
      summary: Replace a segment's name/description/predicate
      tags: [segments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSegmentRequest"
      responses:
        "200":
          description: Segment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Segment"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
    ProjectSlug:
      name: slug
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9][a-z0-9-]{0,62}$"
...
    SegmentKey:
      name: segKey
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9][a-z0-9-_.]{0,127}$"
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: string
...
    Segment:
      type: object
      required: [id, project_id, key, name, description, predicate, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        key:
          type: string
        name:
          type: string
        description:
          type: string
        predicate:
          description: An IR Predicate JSON. Validated server-side at write time.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SegmentList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Segment"
    CreateSegmentRequest:
      type: object
      required: [key, name, predicate]
      properties:
        key:
          type: string
        name:
          type: string
        description:
          type: string
        predicate:
          description: An IR Predicate JSON.
    UpdateSegmentRequest:
      type: object
      required: [name, predicate]
      properties:
        name:
          type: string
        description:
          type: string
        predicate:
          description: An IR Predicate JSON.
...